requirements.txt
django-cors-headers
django
djangorestframework

